renders New Campsite components in screed with conditional logic
